requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.26.0
pandas==1.3.5
python-dotenv==0.19.2
fredapi
pyodbc
azure-identity